@@ -1,8 +1,8 @@
 # encoding: UTF-8
 module SurveyorControllerCustomMethods
   def self.included(base)
-    # base.send :before_filter, :require_user   # AuthLogic
-    # base.send :before_filter, :login_required  # Restful Authentication
+    # base.send :before_action, :require_user   # AuthLogic
+    # base.send :before_action, :login_required  # Restful Authentication
     # base.send :layout, 'surveyor_custom'
   end
 
