@@ -6,12 +6,12 @@ module Surveyor
   module SurveyorControllerMethods
     extend ActiveSupport::Concern
     included do
-      before_filter :get_current_user, :only => [:new, :create]
-      before_filter :determine_if_javascript_is_enabled, :only => [:create, :update]
-      before_filter :set_response_set_and_render_context, :only => [:edit, :show, :delete_response]
+      before_action :get_current_user, :only => [:new, :create]
+      before_action :determine_if_javascript_is_enabled, :only => [:create, :update]
+      before_action :set_response_set_and_render_context, :only => [:edit, :show, :delete_response]
 
       layout 'surveyor_default'
-      before_filter :set_locale
+      before_action :set_locale
     end
 
     # Actions
@@ -39,7 +39,7 @@ module Surveyor
     end
 
     def show
-      # @response_set is set in before_filter - set_response_set_and_render_context
+      # @response_set is set in before_action - set_response_set_and_render_context
       if @response_set
         @survey = @response_set.survey
         respond_to do |format|
@@ -57,7 +57,7 @@ module Surveyor
     end
 
     def edit
-      # @response_set is set in before_filter - set_response_set_and_render_context
+      # @response_set is set in before_action - set_response_set_and_render_context
       if @response_set
         @sections = SurveySection.where(survey_id: @response_set.survey_id).includes([:survey, {questions: [{answers: :question}, {question_group: :dependency}, :dependency]}])
         @section = (section_id_from(params) ? @sections.where(id: section_id_from(params)).first : @sections.first) || @sections.first
@@ -70,7 +70,7 @@ module Surveyor
     end
 
     def update
-      question_ids_for_dependencies = (params[:r] || []).map{|k,v| v["question_id"] }.compact.uniq
+      question_ids_for_dependencies = (params[:r]&.to_unsafe_hash || []).map{|k,v| v["question_id"] }.compact.uniq
       saved = load_and_update_response_set_with_retries
 
       return redirect_with_message(surveyor_finish, :notice, t('surveyor.completed_survey')) if saved && params[:finish]
