@@ -1,3 +1,3 @@
 module Surveyor
-  VERSION = '1.5.2'
+  VERSION = '1.5.3'
 end
